requirements.txt
pyarrow==15.0
pylance 
bitsandbytes 
peft 
trl 
accelerate 
transformers 
langchain 
sentence-transformers 
langchainhub 
langchain_community 
chromadb
langchain-huggingface
pandas
tqdm
datasets
ragas
rank_bm25
python-dotenv